Make Kaggle submissions for class
For class I had to run some simple models on this dataset and then submit them to Kaggle. The accuracy for the decision tree model was 86.9%, and the random forest was 96.5%. It seems kind of silly to me to submit these low scoring models when the first one I made got around 98.5% accuracy, and my current best is somewhere around 99.3%, but these models were required for class for some reason. Probably for comparison, so we can see how much better other models can do? I'm not really sure.